refactor: Simplify VolumeCard component by removing unnecessary props and handlers; add delete functionality to VolumeDetail component
