+++ client/src/pages/VolumeDetail/index.css
@@ -43,3 +43,9 @@
 .volume-detail li {
   margin: 4px 0;
 }
+
+.btn-div{
+  display: flex;
+  justify-content: center;
+  margin-top: 20px;
+}
